+++ contracts/satoshi-bridge/src/bitcoin_utils/mod.rs
@@ -1,2 +1,3 @@
+pub mod contract_methods;
 pub mod psbt_wrapper;
 pub mod transaction;
