+++ contracts/satoshi-bridge/src/lib.rs
@@ -61,10 +61,14 @@ pub use crate::utils::*;
 pub use crate::utxo::*;
 
 #[cfg(not(feature = "zcash"))]
+pub use crate::bitcoin_utils::contract_methods::*;
+#[cfg(not(feature = "zcash"))]
 pub use crate::bitcoin_utils::psbt_wrapper;
 #[cfg(not(feature = "zcash"))]
 pub use crate::bitcoin_utils::transaction::Transaction as WrappedTransaction;
 
+#[cfg(feature = "zcash")]
+pub use crate::zcash_utils::contract_methods::*;
 #[cfg(feature = "zcash")]
 pub use crate::zcash_utils::psbt_wrapper;
 #[cfg(feature = "zcash")]
